pyproject.toml
[tool.poetry]
name = "west-consumer"
version = "0.1.0"
description = "West Consumer"
authors = ["lukaszlacinski <lukasz@uchicago.edu>"]
readme = "README.md"
packages = [
    { include = "src" }
]
package-mode = false

[tool.poetry.dependencies]
boto3 = "^1.35.49"
confluent-kafka = "^2.6.0"
globus-sdk = "^3.46.0"
jsonpatch = "^1.33"
python = "^3.10"
python-dotenv = "^1.0.1"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
